refactor item and order model to change price into DecimalField

+++ catalog/models/order.py
@@ -1,5 +1,7 @@
 from django.db import models
 
+from config.constants import Currency
+
 class Order(models.Model):
     session_key = models.CharField(max_length=40, db_index=True, null=True)
     created_at = models.DateTimeField(auto_now_add=True)
@@ -15,4 +17,7 @@ class OrderItem(models.Model):
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
     item = models.ForeignKey('Item', on_delete=models.CASCADE)
     quantity = models.PositiveIntegerField(default=1)
-    price_at_purchase = models.PositiveIntegerField()
+    price_at_purchase = models.DecimalField(
+        max_digits=10, 
+        decimal_places=2,
+    )
